refactor: organize SCC algorithms into dedicated module
Move kosaraju_scc and tarjan_scc algorithms from algo/mod.rs into separate files under new algo/scc/ module. Maintain backward compatibility through re-exports.
