+++ src/algo/scc/mod.rs
@@ -1,0 +1,5 @@
+pub mod kosaraju_scc;
+pub mod tarjan_scc;
+
+pub use kosaraju_scc::{kosaraju_scc, scc};
+pub use tarjan_scc::{tarjan_scc, TarjanScc};
